pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-tool"
version = "0.1.0"
description = "FastAPI + HTMX shell for a personal multi-agent tool."
readme = "docs/ai/requirements/req-ai-agent-tool.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10,<4.0",
    "fastapi>=0.110,<1.0",
    "filelock>=3.16,<4.0",
    "httpx>=0.27,<1.0",
    "jinja2>=3.1,<4.0",
    "pydantic>=2.7,<3.0",
    "python-multipart>=0.0.9,<1.0",
    "pyyaml>=6.0,<7.0",
    "uvicorn>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
